time library removal
The removal of the time library (more specifically, the sleep function) has been done. Apart from being unnecessary for the inner functioning of the library unlike mitmproxy and json, it accounted for only 4 lines (5 including the function importation) which is very little reliance and usage. This also makes the script speedier and better since 0.25 (or 0.5) seconds have been shaved off the time which is very important to ensure this script is better and improved (and doesn't make the client and server wait). It also makes the server slightly more lightweight!