Uploaded requirements.txt
This file is very important in Python standards since it allows others to install this script's exact libraries and dependencies easily to have the script run, because if this script isn't tailored for a better version of mitmproxy, this file specifies the version. It's also good for general Python usage and teamwork. This file is a upgrade from bombie since it now specifies the third-party libraries (json not included) and their version which makes setup and use easier. This also calls for the removal of the first TODO of the README file.