Fixed minor details of Coldie
The name of the script was revamped to Coldie, the version was stabilized to v0.1-alpha (not v1.0-alpha). The script also gave credits to bombie. It's now ready for a release with all the fancy tids and bits (and these script changes). The new thing is that on each commit, Coldie's commiter will sign off to ensure legitimacy of commit!

Signed-off-by: sandStoop <sandstoop@proton.me>