New feature to Coldie
This update introduces a new feature to Coldie, "Recommended Version" (or "Old Version Bypass"), a feature which tricks the client that the version of the game is the recommended (latest) version in it's API endpoint response. It can be used to play other versions of the game (new features or version-exclusive things won't be there!) This approach is stable and seems to be working, however it has been untested in older versions. A new TODO has been proposed to transform -1 to False to better represent an unset bit with False (0) and certainly not -1 (which evaluates to True and is bad for quite a few reasons).

Signed-off-by: sandStoop <sandstoop@proton.me>