Updated and clarified the README
Added the recommended version "cheat" to the Do not turn off proxy list and explained more about cheats that trick the client and those that trick the server (and clarified the duration they should be on). The most important detail is that the proxy should be enabled for those specific features (and when disabled, the features go however they can enabled back on). It added the new feature and more clarification about the proxy and everything about it. What you need to know about it is that for the features of the proxy, the proxy needs to be on for the permanent features (and when off, they're revoked but they can be brought back) and for the temporary features, they can be on for one time, then off (no problems, no worries).

Signed-off-by: sandStoop <sandstoop@proton.me>