Short correction to README
This included a short correction to the README file since ternary operators (the introduction of them) makes the game crash due to an oversight with the concept of ternary operators which always expects a value (pass is not allowed). Choosing None returns null which makes the game crash since values can't be null (at least BO!). This oversight has been patched before it has made an release with testing and the README has quickly been fixed. In other words, the ternary operator always passes a value/does an operation and never truly nothing.

Signed-off-by: sandStoop <sandstoop@proton.me>