Made the tutorial not be injected on default
This is a one-liner change and part of Coldie's many codebase changes to ensure that the user has full control of the program and unexpected things don't happen. This also ensures that the user can setup the program themselves without unexpected values and more freedom (meaning users will have a higher IQ and follow tutorials better). This change is also used to lead to the eventual usage of the False boolean (unset bit, better for performance and more benefits) and not -1 which will lead to many other changes and QoL improvements.


Signed-off-by: sandStoop <sandstoop@proton.me>