Massive QoL change
bombie's approach to unset bits/falsy values used -1 instead of False. This was hard-to-debug, even harder to setup, isn't the general and it's really hard to decode. This new code uses False instead of -1. The code has also been updated to disable all features by default so that users fine tune it to their every need (anti-ban is still on by default!)

Signed-off-by: sandStoop <sandstoop@proton.me>