Necessary changes to code due to it's independent status
This code has been fully revamped and fixed to give the credits to Coldie and not bombie. bombie has done a lot for the project however it is an inactive repository which is why Coldie is leaving the fork network to be independent of it's old buddy. This code changes comments and logging function strings.

Signed-off-by: sandStoop <sandstoop@proton.me>