Massive QoL update
This is a big update to Coldie which has fixed a ticking time bomb in the code which would have failed some tests. This update makes Coldie slightly shorter (less characters, we're not just hunting for less characters...), more stable and accurate in variable boolean checking, respect to Python's standards (and PEP8) and fixing of weird behavior with strings in a shorter and better way.

Signed-off-by: sandStoop <sandstoop@proton.me>