Added possible drawback to list of drawbacks and talked about certificates
A drawback that has remained since the existence of Bombie has been added, this is not Coldie's fault. It's the faulty domains which do not trust the certificates causing this. They will not trust such certificates sometimes and cause issues. This happens especially for default mitmproxy configurations (even when the certificate is trusted). If untrusted, payments are affected all of the time and OTP codes for e-mail signing are affected most of the time!

Signed-off-by: sandStoop <sandstoop@proton.me>