Replace all usage of the deprecated ctx module with the logging module
According to the mitmproxy API breaking changes changelog, since mitmproxy 9.0 the mitmproxy ctx logging system has been deprecated in favor of the logging module. This ensures that this project won't rely on something that could get removed at any time. It's a solid change to ensure the project relies less on mitmproxy and more on Python's standard library